requirements.txt
alembic==1.14.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
APScheduler==3.10.4
async-timeout==5.0.1
babel==2.16.0
backports.zoneinfo==0.2.1
bandit==1.7.10
blinker==1.8.2
cbor2==5.6.5
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.1
ecdsa==0.19.0
exceptiongroup==1.2.2
flake8==7.1.1
flask==3.0.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy>=3.0.0
Flask-SSE==1.0.0
flask-wtf==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
ic-py==1.0.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
leb128==1.0.8
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mnemonic==0.20
multimethod==1.10
packaging==24.2
pbr==6.1.0
pi==0.1.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pygments==2.18.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stevedore==5.3.0
tomli==2.1.0
typing-extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
waiter==1.4
werkzeug==3.0.6
wtforms==3.1.2
zipp==3.20.2
